refactor setInitial user context
