+++ front_end/src/pages/Login.js
@@ -16,6 +16,7 @@ import { getEpisodeInfo } from "../WebAPI/listenAPI";
 import useInputs from "../hooks/useInputs";
 import useUser from "../hooks/useUser";
 import Input from "../components/UserInput";
+import { setInitialUserContext } from "../utils";
 
 const LoginPageWrapper = styled.div`
   max-width: 1920px;
@@ -159,61 +160,18 @@ export default function Login() {
     } catch (err) {
       console.log(err);
     }
+
     if (result.ok) {
       window.localStorage.removeItem("podcastifyToken");
       window.localStorage.setItem("podcastifyToken", result.token);
-      const response = await getMyInfo(result.token);
-      let {
-        playlists,
-        subscriptions,
-        playedRecords,
-        ...userInfo
-      } = response.data;
-      for (let i = 0; i < playlists.length; i++) {
-        let { Episodes, ...rest } = playlists[i];
-        Episodes = await Promise.all(
-          Episodes.map(async (ep) => {
-            const episodeInfo = await getEpisodeInfo(ep.id);
-            return episodeInfo.data;
-          })
-        );
-        playlists[i] = { Episodes, ...rest };
-      }
-
-      // 節省打 API 次數，只取最後三筆播放紀錄
-      let lastThreePlayedRecords = [];
-      if (playedRecords.length > 3) {
-        for (let i = 0; i < 3; i++) {
-          lastThreePlayedRecords[i] = playedRecords[i];
-        }
-      } else {
-        for (let i = 0; i < playedRecords.length; i++) {
-          lastThreePlayedRecords[i] = playedRecords[i];
-        }
-      }
-
-      // 拿到播放紀錄的單集詳細資料
-      let playedRecordsDetails = await Promise.all(
-        lastThreePlayedRecords.map(async (ep) => {
-          if (ep.episodeId.length !== 32 || ep.progress === 0) return;
-          const episodeInfo = await getEpisodeInfo(ep.episodeId);
-          return episodeInfo.data;
-        })
+      setInitialUserContext(
+        getMyInfo,
+        getEpisodeInfo,
+        setUserInfo,
+        setUserPlaylists,
+        setUserPlayedRecord,
+        setUserSubscription
       );
-
-      // 播放紀錄資料重整
-      let record = [];
-      for (let i = 0; i < playedRecordsDetails.length; i++) {
-        record[i] = {
-          episode: playedRecordsDetails[i],
-          progress: lastThreePlayedRecords[i].progress,
-        };
-      }
-
-      setUserInfo(userInfo);
-      setUserPlaylists(playlists);
-      setUserSubscription(subscriptions);
-      setUserPlayedRecord(record);
       history.push("/");
     } else {
       window.localStorage.removeItem("podcastifyToken");
